+++ jrnl/util.py
@@ -3,15 +3,36 @@
 import sys
 import os
 from tzlocal import get_localzone
+import getpass as gp
 
+PY3 = sys.version_info[0] == 3
+PY2 = sys.version_info[0] == 2
+STDIN = sys.stdin
+STDERR = sys.stderr
+STDOUT = sys.stdout
+TEST = False
 __cached_tz = None
 
-def py23_input(msg):
-    if sys.version_info[0] == 3:
-        try: return input(msg)
-        except SyntaxError: return ""
+def getpass(prompt):
+    if not TEST:
+        return gp.getpass(prompt)
     else:
-        return raw_input(msg)
+        return py23_input(prompt)
+
+
+def u(s):
+    """Mock unicode function for python 2 and 3 compatibility."""
+    return s if PY3 or type(s) is unicode else unicode(s, "unicode_escape")
+
+def prompt(msg):
+    """Prints a message to the std err stream defined in util."""
+    if not msg.endswith("\n"):
+        msg += "\n"
+    STDERR.write(u(msg))
+
+def py23_input(msg):
+    STDERR.write(u(msg))
+    return STDIN.readline().strip()
 
 def get_local_timezone():
     """Returns the Olson identifier of the local timezone.